remove old search and rename SearchNew to Search
