+++ src/components/Search/CloseButton/CloseButton.js
