+++ src/components/Search/CloseButton/index.js
