--- src/components/search/search.scss
@@ -1,81 +1,0 @@
-.search-results-list {
-    position: absolute;
-    top: 0;
-    right: 0;
-    width: 800px;
-    display: none;
-    &.on {
-        display: block;
-    }
-    .title {
-        font-family: '72-Regular';
-        font-size: 18px;
-        color: #C1C0C0;
-        line-height: 70px;
-    }
-
-    .search-categories {
-        height: 50px;
-        div {
-            line-height: 50px;
-            float: left;
-            width: 200px;
-            font-family: '72-Light';
-            font-size: 20px;
-            cursor: pointer;
-            color: #9D9D9D;
-            &.selected {
-                cursor: default;
-                color: #E3E3E3;
-            }
-            &.close {
-                float: right;
-                margin-right: 33px;
-                margin-top: 18px;
-                width: 16px;
-                height: 16px;
-                cursor: pointer;
-                bottom: 0px;
-                background-image: url('img/cross.png');
-            }
-        }
-
-    }
-    .search-results-wrapper {
-        height: 606px;
-        overflow-y: auto;
-        padding: 7px;
-        overflow-x: initial;
-        .search-result {
-
-
-            min-height: 200px;
-            padding-left: 200px;
-            padding-right: 100px;
-
-            .title {
-                font-family: '72-Bold';
-                font-size: 24px;
-                color: #C9C9C9;
-                line-height: initial;
-                padding-top: 20px;
-            }
-            .description {
-                font-family: '72-Light';
-                font-size: 18px;
-                color: #BDBDBD;
-            }
-            .search-section {
-                font-family: '72-Regular';
-                font-size: 14px;
-                color: #ACC1F3;
-                margin-top: 5px;
-                padding-bottom: 10px;
-            }
-
-            &:hover {
-                box-shadow: 1px 1px 10px 1px black;
-            }
-        }
-    }
-  }
